endpoints: finishing/testing all Controllers endpoints ✅

+++ src/main/java/com/peacock/backend/dto/BillItemDTO.java
@@ -1,4 +1,4 @@
 package com.peacock.backend.dto;
 
-public record BillItemDTO(Long product_id, Long quantity) {
+public record BillItemDTO(Long product_id, Integer quantity) {
 }
